requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==3.4.8
cycler==0.11.0
ecdsa==0.17.0
fastapi==0.78.0
fonttools==4.33.3
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
